API

Aircraft Geometry API

AeroMDAO.AircraftGeometry.AbstractSpacing — Type

An abstract type to define custom spacing distributions.

AeroMDAO.AircraftGeometry.Foil — Type

Foil(x, y, name = "Unnamed")
Foil(coordinates, name = "Unnamed")

Structure consisting of foil coordinates in 2 dimensions with an optional name.

The coordinates should be provided in Selig format for compatibility with other AeroMDAO tools.

AeroMDAO.AircraftGeometry.HalfWing — Type

HalfWing(; foils :: Vector{Foil}, 
           chords, 
           twists, 
           spans, 
           dihedrals, 
           LE_sweeps,
           position = zeros(3),
           angle    = 0.
           axis     = [0.,1.,0.])

Definition for a HalfWing consisting of $N+1$ Foils, their associated chord lengths $c$ and twist angles $ι$, for $N$ sections with span lengths $b$, dihedrals $δ$ and sweep angles $Λ$, with all angles in degrees.

AeroMDAO.AircraftGeometry.Wing — Type

Wing(; foils, chords, twists, 
       spans, dihedrals, LE_sweeps,
       position, angle, axis)

Wing(left :: HalfWing, right :: HalfWing)
Wing(half :: HalfWing)

A composite type consisting of fields left and right, each a HalfWing for constructing a wing.

A single argument generates a symmetric Wing.

Arguments

foils :: Vector{Foil}

chords

twists

spans

dihedrals

LE_sweeps

position = zeros(3)

angle = 0

axis = [0.,1.,0.]

AeroMDAO.AircraftGeometry.HalfWingSection — Method

HalfWingSection(; span, dihedral, LE_sweep, taper, root_chord,
                  root_twist, tip_twist, root_foil, tip_foil,
                  position, angle, axis)

Define a HalfWing consisting of a single trapezoidal section.

Arguments

span :: Real = 1.: Span length

dihedral :: Real = 1.: Dihedral angle (degrees)

LE_sweep :: Real = 0.: Leading-edge sweep angle (degrees)

taper :: Real = 1.: Taper ratio of tip to root chord

root_chord :: Real = 1.: Root chord length

root_twist :: Real = 0.: Twist angle at root (degrees)

tip_twist :: Real = 0.: Twist angle at tip (degrees)

root_foil :: Array{Real} = naca4((0,0,1,2)): Foil coordinates at root

tip_foil :: Array{Real} = naca4((0,0,1,2)): Foil coordinates at tip

position :: Vector{Real} = zeros(3): Position

angle :: Real = 0.: Angle of rotation (degrees)

axis :: Vector{Real} = [0.,1.,0.]: Axis of rotation

AeroMDAO.AircraftGeometry.WingSection — Method

WingSection(; span, dihedral, LE_sweep, taper, root_chord,
              root_twist, tip_twist, root_foil, tip_foil,
              position, angle, axis)

Define a Wing consisting of two trapezoidal sections with one plane of symmetry.

Arguments

span :: Real = 1.: Span length

dihedral :: Real = 1.: Dihedral angle (degrees)

LE_sweep :: Real = 0.: Leading-edge sweep angle (degrees)

taper :: Real = 1.: Taper ratio of tip to root chord

root_chord :: Real = 1.: Root chord length

root_twist :: Real = 0.: Twist angle at root (degrees)

tip_twist :: Real = 0.: Twist angle at tip (degrees)

root_foil :: Array{Real} = naca4((0,0,1,2)): Foil coordinates at root

tip_foil :: Array{Real} = naca4((0,0,1,2)): Foil coordinates at tip

position :: Vector{Real} = zeros(3): Position

angle :: Real = 0.: Angle of rotation (degrees)

axis :: Vector{Real} = [0.,1.,0.]: Axis of rotation

AeroMDAO.AircraftGeometry.arc_length — Method

arc_length(foil :: Foil)

Compute the arc-length of a Foil.

AeroMDAO.AircraftGeometry.aspect_ratio — Method

aspect_ratio(wing :: AbstractWing)

Compute the aspect ratio of an AbstractWing.

AeroMDAO.AircraftGeometry.camber_CST — Function

camber_CST(α_c, α_t,
           (Δz_u, Δz_l) :: NTuple{2, Real},
           coeff_LE = 0.,
           n :: Integer = 40)

Define a cosine-spaced foil with $2n$ points using the Class Shape Transformation method on a Bernstein polynomial basis for the camber and thickness coordinates.

The foil is defined by arrays of coefficients $(α_c,~ α_t)$ for the upper and lower surfaces, trailing-edge spacing values $(Δz_u,~Δz_l)$, and a coefficient for the leading edge modifications.

AeroMDAO.AircraftGeometry.camber_coordinates — Method

camber_coordinates(coords :: Array{2, <: Real})

Generate the camber coordinates on the $x$-$z$ plane at $y = 0$.

AeroMDAO.AircraftGeometry.camber_coordinates — Method

camber_coordinates(wing :: HalfWing, n_s :: Integer, n_c :: Integer, flip = false)

Compute the camber coordinates of a HalfWing consisting of camber distributions of Foils and relevant geometric quantities, given numbers of spanwise $n_s$ and chordwise $n_c$ panels, with an option to flip the signs of the $y$-coordinates.

AeroMDAO.AircraftGeometry.camber_thickness — Function

camber_thickness(foil :: Foil, num :: Integer)

Compute the camber-thickness distribution of a Foil with cosine spacing.

AeroMDAO.AircraftGeometry.camber_thickness_to_coordinates — Method

camber_thickness_to_coordinates(xs, camber, thickness)

Convert the camber-thickness representation to 2-dimensional coordinates given the $x$-locations and their corresponding camber and thickness values.

AeroMDAO.AircraftGeometry.chord_coordinates — Method

chord_coordinates(wing :: AbstractWing, n_s :: Integer, n_c :: Integer, flip = false)

Compute the chord coordinates of a HalfWing consisting of Foils and relevant geometric quantities, given numbers of spanwise $n_s$ and chordwise $n_c$ panels, with an option to flip the signs of the $y$-coordinates.

AeroMDAO.AircraftGeometry.coordinates — Method

coordinates(foil :: Foil)

Generate the array of Foil coordinates.

AeroMDAO.AircraftGeometry.coordinates_to_camber_thickness — Function

coordinates_to_camber_thickness(coords, n = 40)

Convert 2-dimensional coordinates to its camber-thickness representation after cosine interpolation with $2n$ points.

AeroMDAO.AircraftGeometry.cosine_foil — Function

cosine_foil(coordinates, n = 40)

Interpolate a foil profile's coordinates by projecting the x-coordinates of a circle onto the geometry with $2n$ points.

AeroMDAO.AircraftGeometry.cosine_foil — Method

cosine_foil(foil :: Foil, num :: Integer)

Return a Foil with cosine spacing for a given number of points.

AeroMDAO.AircraftGeometry.kulfan_CST — Function

kulfan_CST(alpha_u, alpha_l,
           (Δz_u, Δz_l) = (0., 0.),
           (LE_u, LE_l) = (0., 0.),
           n            = 40)

Define a cosine-spaced foil with $2n$ points using the Class Shape Transformation method on a Bernstein polynomial basis for the upper and lower coordinates.

The foil is defined by arrays of coefficients $(α_u,~ α_l)$ for the upper and lower surfaces, trailing-edge spacing values $(Δz_u,~ Δz_l)$, and coefficients for leading edge modifications.

AeroMDAO.AircraftGeometry.leading_edge — Function

leading_edge(wing :: HalfWing, flip = false)

Compute the leading edge coordinates of a HalfWing, with an option to flip the signs of the $y$-coordinates.

AeroMDAO.AircraftGeometry.mean_aerodynamic_center — Method

mean_aerodynamic_chord(wing :: AbstractWing)

Compute the coordinates of the mean aerodynamic center of an AbstractWing.

AeroMDAO.AircraftGeometry.mean_aerodynamic_chord — Method

mean_aerodynamic_chord(wing :: AbstractWing)

Compute the mean aerodynamic chord of an AbstractWing.

AeroMDAO.AircraftGeometry.mesh_cambers — Method

mesh_cambers(wing :: AbstractWing, n_s :: Integer, n_c :: Integer; spacings = symmetric_spacing(wing))

Mesh the camber distribution of a Wing into panels of $n_s$ spanwise divisions per section and $n_c$ chordwise divisions with an AbstractSpacing` distribution.

AeroMDAO.AircraftGeometry.mesh_horseshoes — Method

mesh_horseshoes(wing :: AbstractWing, n_s :: Vector{Integer}, n_c :: Integer; flip = false)

Mesh the span and chord distributions of an AbstractWing with $n_s$ spanwise divisions per section and $n_c$ chordwise divisions.

AeroMDAO.AircraftGeometry.mesh_wing — Method

mesh_horseshoes(wing :: AbstractWing, n_s :: Vector{Integer}, n_c :: Integer; flip = false)

Mesh the span and airfoil coordinate distributions of an AbstractWing with $n_s$ spanwise divisions per section and $n_c$ chordwise divisions.

AeroMDAO.AircraftGeometry.naca4 — Function

naca4(digits :: NTuple{4, <: Real}, n :: Integer; sharp_trailing_edge :: Bool)

Generate the coordinates of a NACA 4-digit series profile.

AeroMDAO.AircraftGeometry.projected_area — Method

projected_area(wing :: AbstractWing)

Compute the projected area of an AbstractWing` by summing the trapezoidal areas.

AeroMDAO.AircraftGeometry.read_foil — Method

read_foil(path :: String; header = true)

Read a '.dat' file consisting of 2D coordinates, for an airfoil as an array of SVectors, with an optional argument to skip the header.

AeroMDAO.AircraftGeometry.scale_foil — Method

scale_foil(foil :: Foil, scale)

Scale the coordinates of a Foil to a scaling value.

AeroMDAO.AircraftGeometry.span — Method

span(wing :: AbstractWing)

Compute the planform span of an AbstractWing.

AeroMDAO.AircraftGeometry.surface_coordinates — Method

surface_coordinates(wing :: HalfWing, n_s :: Integer, n_c :: Integer, flip = false)

Compute the surface coordinates of a HalfWing consisting of Foils and relevant geometric quantities, given numbers of spanwise $n_s$ and chordwise $n_c$ panels, with an option to flip the signs of the $y$-coordinates.

AeroMDAO.AircraftGeometry.thickness_coordinates — Method

thickness_coordinates(coords :: Array{2, <: Real})

Generate the thickness coordinates on the $x$-$z$ plane at $y = 0$.

AeroMDAO.AircraftGeometry.wing_bounds — Function

wing_bounds(wing :: HalfWing, flip = false)

Compute the leading and trailing edge coordinates of a HalfWing, with an option to flip the signs of the $y$-coordinates.

AeroMDAO.AircraftGeometry.wing_bounds — Method

wing_bounds(wing :: Wing)

Return the leading and trailing edge coordinates of a Wing.

AeroMDAO.properties — Method

properties(wing :: AbstractWing)

Compute the generic properties of interest (span, area, etc.) of an AbstractWing.

Panel Geometry API

AeroMDAO.PanelGeometry.Panel3D — Type

Panel3D(p1, p2, p3, p4)

Four Cartesian coordinates p1, p2, p3, p4 representing corners of a panel in 3 dimensions. The following commutative diagram (math joke) depicts the order:

z → y
↓
x
        p1 —→— p4
        |       |
        ↓       ↓
        |       |
        p2 —→— p3

AeroMDAO.PanelGeometry.make_panels — Method

make_panels(xyzs)

Convert an array of coordinates corresponding to a wing, ordered from root to tip and leading-edge to trailing-edge, into panels.

AeroMDAO.PanelGeometry.midpoint — Method

midpoint(panel :: Panel3D)

Compute the midpoint of a Panel3D.

AeroMDAO.PanelGeometry.panel_area — Method

panel_area(panel :: Panel3D)

Compute the (possibly non-planar, hence nonsensical) area of a Panel3D.

AeroMDAO.PanelGeometry.panel_coordinates — Method

panel_coordinates(panel :: Panel3D)

Compute the coordinates of a Panel3D.

AeroMDAO.PanelGeometry.panel_normal — Method

panel_normal(panel :: Panel3D)

Compute the normal vector of a Panel3D.

AeroMDAO.PanelGeometry.reflect_xz — Method

reflect_xz(panel :: Panel3D)

Reflect a Panel3D with respect to the $x$-$z$ plane of its reference coordinate system.

AeroMDAO.PanelGeometry.transform — Method

transform(panel :: Panel3D, rotation, translation)

Perform an affine transformation on the coordinates of a Panel3D given a rotation matrix and translation vector.

AeroMDAO.PanelGeometry.transform_normal — Method

transform_normal(panel :: Panel3D, h_l, g_l)

Transform the normal vector of a Panel3D about the hinge axis $ĥ_l$ by the control gain $g_l$.

AeroMDAO.PanelGeometry.wetted_area — Method

wetted_area(panels :: Array{Panel3D})

Compute the total wetted area by summing the areas of an array of Panel3D.

Doublet-Source Panel API

AeroMDAO.DoubletSource.boundary_vector — Method

boundary_vector(panels, u)

Create the vector for the boundary condition of the problem given an array of Panel2Ds and velocity $u$.

AeroMDAO.DoubletSource.doublet_matrix — Method

doublet_matrix(panels_1, panels_2)

Create the matrix of doublet potential influence coefficients between pairs of panels1 and panels2.

AeroMDAO.DoubletSource.eval_coefficients — Method

aerodynamic_coefficients(vels, Δrs, panel_angles, speed, α)

Compute the lift, moment, and pressure coefficients given associated arrays of edge speeds, adjacent collocation point distances, panel angles, the freestream speed, and angle of attack $α$.

AeroMDAO.DoubletSource.influence_matrix — Method

influence_matrix(panels, wake_panel)

Assemble the Aerodynamic Influence Coefficient matrix consisting of the doublet matrix, wake vector, Kutta condition given Panel2Ds and the wake panel.

AeroMDAO.DoubletSource.kutta_condition — Method

kutta_condition(panels)

Create the vector describing Morino's Kutta condition given Panel2Ds.

AeroMDAO.DoubletSource.solve_strengths — Method

solve_strengths(panels, u, wakes, bound)

Solve the system of equations $[AIC][\phi] = [\vec{U} \cdot \hat{n}] - B[\sigma]$ condition given the array of Panel2Ds, a velocity $\vec U$, a vector of wake Panel2Ds, and an optional named bound for the length of the wake.

AeroMDAO.DoubletSource.solve_strengths — Method

solve_strengths(panels, u, sources, bound)

Solve the system of equations $[AIC][\phi] = [\vec{U} \cdot \hat{n}] - B[\sigma]$ condition given the array of Panel2Ds, a velocity $\vec U$, a condition whether to disable source terms ($σ = 0$), and an optional named bound for the length of the wake.

AeroMDAO.DoubletSource.source_matrix — Method

source_matrix(panels_1, panels_2)

Create the matrix of source potential influence coefficients between pairs of panels_1 and panels_2.

AeroMDAO.DoubletSource.source_strengths — Method

source_strengths(panels, freestream)

Create the vector of source strengths for the Dirichlet boundary condition $σ = \vec U_{\infty} \cdot \hat{n}$ given Panel2Ds and a Uniform2D.

AeroMDAO.DoubletSource.tangential_velocities — Method

tangential_velocities(panels, φs, u, sources :: Bool)

Compute the tangential velocities and panel distances given the array of Panel2Ds, their associated doublet strengths $φ$s, the velocity $u$, and a condition whether to disable source terms ($σ = 0$).

AeroMDAO.DoubletSource.wake_vector — Method

wake_vector(wake_panel, panels)

Create the vector of doublet potential influence coefficients from the wake on the panels given the wake panel and the array of Panel2Ds.

Vortex Lattice API

AeroMDAO.VortexLattice.Freestream — Type

Freestream(α, β, Ω)
Freestream(α, β, Ω_x, Ω_y, Ω_z)
Freestream(U, Ω)
Freestream(; 
             alpha = 0., 
             beta  = 0., 
             omega = [0,0,0]
           )

Define freestream conditions with angle of attack $α$ (degrees), sideslip angle $β$ (degrees), and (quasi-steady) rotation vector $Ω$ for a vortex lattice analysis.

Alternatively, provide the velocity vector $U$, which is normalized to determine the angles.

AeroMDAO.VortexLattice.Horseshoe — Type

Horseshoe(panel :: Panel3D, normal, drift = zeros(3))

Generate a Horseshoe corresponding to a Panel3D, an associated normal vector, and a "drift velocity".

AeroMDAO.VortexLattice.Horseshoe — Type

Horseshoe(bound_leg, collocation_point, normal, chord)

Define a horseshoe vortex based on a bound leg, a collocation point, a normal vector, and a "chord length".

AeroMDAO.VortexLattice.Line — Type

Line(r1 :: SVector{3,<: Real}, r2 :: SVector{3,<: Real})

Define a line segment based on two 3-dimensional vectors.

AeroMDAO.VortexLattice.References — Type

References(V, ρ, μ, S, b, c, r)
References(; speed, density, viscosity,
             sound_speed, area, span, 
             chord, location)

Define reference values with speed $V$, density $ρ$, dynamic viscosity $μ$, area $S$, span $b$, chord $c$, location $r$ for a vortex lattice analysis.

Arguments

speed :: Real = 1.: Speed (m/s)

density :: Real = 1.225: Density (m)

viscosity :: Real = 1.5e-5: Dynamic viscosity (m)

sound_speed :: Real = 330.: Speed of sound (m/s)

span :: Real = 1.: Span length (m)

area :: Real = 1.: Area (m²)

chord :: Real = 1.: Chord length (m)

location :: Vector{Real} = [0,0,0]: Position

AeroMDAO.VortexLattice.VortexRing — Type

VortexRing{T <: Real}(left_leg, bound_leg, back_leg, right_leg)

A vortex ring type consisting of four finite vortex line filaments connected in a loop.

AeroMDAO.VortexLattice.VortexRing — Method

Constructor for vortex rings on a Panel3D using Lines. The following convention is adopted:

    p1 —bound_leg→ p4
    |               |
left_leg       right_leg
    ↓               ↓
    p2 —back_leg→ p3

AeroMDAO.VortexLattice.body_frame_velocity — Method

body_frame_velocity(freestream :: Freestream)

Compute the velocity of Freestream in the body reference frame.

AeroMDAO.VortexLattice.bound_leg — Method

bound_leg(horseshoe :: Horseshoe)

Getter for bound leg field of a Horseshoe.

AeroMDAO.VortexLattice.bound_leg — Method

bound_leg(panel :: Panel3D)

Compute the bound leg for a Panel3D, for horseshoes/vortex rings.

AeroMDAO.VortexLattice.bound_leg_center — Method

Compute the midpoint of the bound leg of a Horseshoe.

AeroMDAO.VortexLattice.bound_leg_vector — Method

Compute the direction vector of the bound leg of a Horseshoe.

AeroMDAO.VortexLattice.boundary_condition — Method

boundary_condition(velocities, normals)

Compute the projection of a velocity vector with respect to normal vectors of panels. Corresponds to construction of the boundary condition for the RHS of the AIC system.

AeroMDAO.VortexLattice.compute_farfield_forces — Method

compute_farfield_forces(Δφs, Δs, ∂φ_∂n, θs, V, ρ)

Compute the aerodynamic forces in the Trefftz plane given cumulative doublet strengths $Δφ$s, Trefftz panel lengths $Δs$, doublet-normal directional derivatives $∂φ/∂n$, Trefftz panel angles $θ$s, the freestream speed $V$ and density $ρ$.

AeroMDAO.VortexLattice.doublet_normal_derivatives — Method

doublet_normal_derivatives(wake_lines :: Vector{<: Line}, Δφs, normals)

Compute the normal derivative strengths of the doublets given the wake Lines, net doublet strengths $Δφ$s, and associated normal vectors.

AeroMDAO.VortexLattice.farfield_forces — Method

farfield_forces(Γs, horseshoes, freestream, ρ)

Compute the aerodynamic forces in the Trefftz plane normal to the freestream given horseshoes, their associated strengths $Γ$s, and a density $ρ$.

AeroMDAO.VortexLattice.farfield_influence_matrix — Method

farfield_influence_matrix(centers, normals, points)

Compute the aerodynamic influence coefficient matrix of the wake in the Trefftz plane given the center points, normal vectors, and the points of the wake panels.

AeroMDAO.VortexLattice.farfield_velocity — Method

farfield_velocity(r_i, r_j, Γ_j)

Compute the induced velocity at point $r_i$ of the wake in the Trefftz plane due to the strength $Γ_j$ at $r_j$.

AeroMDAO.VortexLattice.flip_xz — Method

flip_xz(vector)

Reflect the $x$- and $z$- coordinates of a given 3-dimensional vector about the $y$-$z$ and $x$-$y$ planes respectively for the representation in body axes.

AeroMDAO.VortexLattice.geometry_to_stability_axes — Method

geometry_to_stability_axes(coords, α)

Convert coordinates from geometry to stability axes with angle $α$.

AeroMDAO.VortexLattice.geometry_to_wind_axes — Method

geometry_to_wind_axes(coords, α, β)
geometry_to_wind_axes(line :: Line, α, β)

Convert coordinates from geometry axes to wind axes for given angles of attack $α$ and sideslip \beta.``

AeroMDAO.VortexLattice.horseshoe_point — Method

collocation_point(panel :: Panel3D)

Compute the collocation point of a Panel3D for horseshoes/vortex rings, which is the 3-quarter point on each side in the $x$-$z$ plane.

AeroMDAO.VortexLattice.influence_coefficient — Function

influence_coefficient(r, horseshoe, normal, V_hat, symmetry)

Compute the influence coefficient of the velocity of a Horseshoe with trailing lines in a given direction $û$ at a point $r$ projected to a normal vector.

AeroMDAO.VortexLattice.influence_matrix — Function

influence_matrix(horseshoes, collocation_points, normals, V_hat)

Assemble the Aerodynamic Influence Coefficient (AIC) matrix given horseshoes, collocation points, associated normal vectors, and a unit vector representing the freestream.

AeroMDAO.VortexLattice.project_yz — Method

project_yz(vector)
project_yz(line :: Line)

Project a given 3-dimensional vector or Line into the $y$-$z$ plane.

AeroMDAO.VortexLattice.reflect_xz — Method

reflect_xz(vector)
reflect_xz(line :: Line)
reflect_xz(horseshoe :: Horseshoe)

Reflect the $y$-coordinate of a given 3-dimensional vector (or $Line$ or $Horseshoe$) about the $x$-$z$ plane.

AeroMDAO.VortexLattice.stability_to_geometry_axes — Method

geometry_to_stability_axes(coords, α)

Convert coordinates from stability to geometry axes with angle $α$.

AeroMDAO.VortexLattice.stream_velocity — Method

stream_velocity(r, horseshoes, Γs, V, Ω)

Evaluate the total induced velocity at a point r given Horseshoes, vortex strengths Γs, rotation rates Ω, and freestream flow vector freestream in the global reference frame.

AeroMDAO.VortexLattice.streamlines — Method

streamlines(point, freestream :: Freestream, horseshoes, Γs, length, num_steps)

Compute the streamlines from a given starting point, a Freestream, Horseshoes and their associated strengths Γs with a specified length of the streamline and number of evaluation points.

AeroMDAO.VortexLattice.sum_vortices — Method

sum_vortices(r, vortex_lines :: Array{Line}, Γ)

Sums the velocities evaluated at a point $r$ of Lines with constant strength $Γ$.

AeroMDAO.VortexLattice.surface_dynamics — Method

surface_dynamics(system :: VLMSystem; 
                 axes   :: AbstractAxisSystem = Geometry())

Compute the forces and moments on the surface given the VLMSystem after performing an analysis, and the reference axis system.

AeroMDAO.VortexLattice.surface_forces — Method

surface_forces(hs_comp, Γ_comp, horseshoes, Γs, U, Ω, ρ)

Compute the surface forces via the local Kutta-Jowkowsky theorem.

For a given array of vortex type hs_comp and their associated vortex strengths $Γ_c$ for which to compute the forces, the arrays of horseshoes and vortex strengths $Γ$s of the entire aircraft, the freestream flow vector $U$, rotation rates $Ω$, and a density $ρ$, the velocities are evaluated at the midpoint of the bound leg of each horseshoe, excluding the contribution of the bound leg.

AeroMDAO.VortexLattice.surface_forces — Method

surface_forces(system :: VLMSystem; 
               axes   :: AbstractAxisSystem = Geometry())

Compute the forces on the surface given the VLMSystem after performing an analysis, and the reference axis system.

AeroMDAO.VortexLattice.surface_moments — Method

surface_moments(system :: VLMSystem; 
                axes   :: AbstractAxisSystem = Geometry())

Compute the moments on the surface given the VLMSystem after performing an analysis, and the reference axis system.

AeroMDAO.VortexLattice.surface_velocities — Method

surface_velocities(system :: VLMSystem; 
                   axes   :: AbstractAxisSystem = Geometry())

Compute the velocities on the surface given the VLMSystem after performing an analysis, and the reference axis system.

AeroMDAO.VortexLattice.trailing_velocity — Method

trailing_velocity(r, horseshoe, Γ, V_hat)

Compute the induced velocities at a point $r$ of a given Horseshoe with constant strength $Γ$ and trailing legs pointing in a given direction $û$, excluding the bound leg contribution.

AeroMDAO.VortexLattice.trefftz_plane_quantities — Method

trefftz_plane_quantities(horseshoes :: AbstractArray{<: Horseshoe}, α, β)

Project Horseshoes into the Trefftz plane aligned with the wind axes angles $α,~β$, and compute normal vectors, projection angles and lengths.

AeroMDAO.VortexLattice.vortex_ring — Method

Helper function to compute the vortex ring given four points following Panel3D ordering.

AeroMDAO.VortexLattice.vortex_ring — Method

Computes the vortex rings on a Panel3D.

AeroMDAO.VortexLattice.wind_to_geometry_axes — Method

wind_to_geometry_axes(coords, α, β)

Convert coordinates from wind axes to geometry axes for given angles of attack $α$ and sideslip \beta.``

AeroMDAO.solve_linear — Function

solve_linear(horseshoes, normals, U, Ω)

Evaluate and return the vortex strengths $Γ$s given Horseshoes, their associated normal vectors (not necessarily the same as the panels' normals), the speed $U$ and rotation vector $Ω$.

AeroMDAO.velocity — Function

velocity(r, horseshoe, Γ, V_hat)

Compute the induced velocities at a point $r$ of a given Horseshoe with constant strength $Γ$ and trailing legs pointing in a given direction $û$.

AeroMDAO.velocity — Method

Computes the induced velocities at a point $r$ of a VortexRing with constant strength $Γ$.

AeroMDAO.velocity — Method

velocity(freestream :: Freestream)

Compute the velocity of a Freestream.

Beams API

AeroMDAO.Beams.Material — Method

Material(E, J, σ_max, ρ)
Material(; elastic_modulus, shear_modulus, yield_stress, density)

Define a Material with positive, real elastic modulus $E$, shear modulus $G$, yield stress $σ_{\max}$, and density $ρ$.

The default assignments for the named variables are set to the properties of aluminium?

Arguments

elastic_modulus :: Real = 85e9: Elastic modulus.

shear_modulus :: Real = 25e9: Shear modulus.

yield_stress :: Real = 350e6: Yield stress.

density :: Real = 1.6e3: Density.

AeroMDAO.Beams.Tube — Type

Tube(material :: Material, length, radius, thickness)

Define a hollow tube of fixed radius with a given material, length, and thickness.

AeroMDAO.Beams.tube_stiffness_matrix — Function

tube_stiffness_matrix(E, G, A, Iyy, Izz, J, L)
tube_stiffness_matrix(E, G, A, Iyy, Izz, J, L, num)
tube_stiffness_matrix(tube :: Tube)
tube_stiffness_matrix(tube :: Vector{Tube})
tube_stiffness_matrix(x :: Matrix{Real})

Generate the stiffness matrix using the properties of a tube.

The required properties are the elastic modulus $E$, shear modulus $G$, area $A$, moments of inertia about the $y-$ and $z-$ axes $I_{yy}, I_{zz}$, polar moment of inertia $J$, length $L$. A composite stiffness matrix is generated with a specified number of elements.

Aerostructural API