Geometry API

Aircraft Geometry API

AeroMDAO.AircraftGeometry.AbstractSpacing — Type

An abstract type to define custom spacing distributions.

AeroMDAO.AircraftGeometry.Foil — Type

Foil(x, y, name = "Unnamed")
Foil(coordinates, name = "Unnamed")

Structure consisting of foil coordinates in 2 dimensions with an optional name.

The coordinates should be provided in Selig format for compatibility with other AeroMDAO tools.

AeroMDAO.AircraftGeometry.HalfWing — Type

HalfWing(; foils :: Vector{Foil}, 
           chords, 
           twists, 
           spans, 
           dihedrals, 
           LE_sweeps,
           position = zeros(3),
           angle    = 0.
           axis     = [0.,1.,0.])

Definition for a HalfWing consisting of $N+1$ Foils, their associated chord lengths $c$ and twist angles $ι$, for $N$ sections with span lengths $b$, dihedrals $δ$ and sweep angles $Λ$, with all angles in degrees.

AeroMDAO.AircraftGeometry.Wing — Type

Wing(; foils, chords, twists, 
       spans, dihedrals, LE_sweeps,
       position, angle, axis)

Wing(left :: HalfWing, right :: HalfWing)
Wing(half :: HalfWing)

A composite type consisting of fields left and right, each a HalfWing for constructing a wing.

A single argument generates a symmetric Wing.

Arguments

foils :: Vector{Foil}

chords

twists

spans

dihedrals

LE_sweeps

position = zeros(3)

angle = 0

axis = [0.,1.,0.]

AeroMDAO.AircraftGeometry.HalfWingSection — Method

HalfWingSection(; span, dihedral, LE_sweep, taper, root_chord,
                  root_twist, tip_twist, root_foil, tip_foil,
                  position, angle, axis)

Define a HalfWing consisting of a single trapezoidal section.

Arguments

span :: Real = 1.: Span length

dihedral :: Real = 1.: Dihedral angle (degrees)

LE_sweep :: Real = 0.: Leading-edge sweep angle (degrees)

taper :: Real = 1.: Taper ratio of tip to root chord

root_chord :: Real = 1.: Root chord length

root_twist :: Real = 0.: Twist angle at root (degrees)

tip_twist :: Real = 0.: Twist angle at tip (degrees)

root_foil :: Array{Real} = naca4((0,0,1,2)): Foil coordinates at root

tip_foil :: Array{Real} = naca4((0,0,1,2)): Foil coordinates at tip

position :: Vector{Real} = zeros(3): Position

angle :: Real = 0.: Angle of rotation (degrees)

axis :: Vector{Real} = [0.,1.,0.]: Axis of rotation

AeroMDAO.AircraftGeometry.WingSection — Method

WingSection(; span, dihedral, LE_sweep, taper, root_chord,
              root_twist, tip_twist, root_foil, tip_foil,
              position, angle, axis)

Define a Wing consisting of two trapezoidal sections with one plane of symmetry.

Arguments

span :: Real = 1.: Span length

dihedral :: Real = 1.: Dihedral angle (degrees)

LE_sweep :: Real = 0.: Leading-edge sweep angle (degrees)

taper :: Real = 1.: Taper ratio of tip to root chord

root_chord :: Real = 1.: Root chord length

root_twist :: Real = 0.: Twist angle at root (degrees)

tip_twist :: Real = 0.: Twist angle at tip (degrees)

root_foil :: Array{Real} = naca4((0,0,1,2)): Foil coordinates at root

tip_foil :: Array{Real} = naca4((0,0,1,2)): Foil coordinates at tip

position :: Vector{Real} = zeros(3): Position

angle :: Real = 0.: Angle of rotation (degrees)

axis :: Vector{Real} = [0.,1.,0.]: Axis of rotation

AeroMDAO.AircraftGeometry.arc_length — Method

arc_length(foil :: Foil)

Compute the arc-length of a Foil.

AeroMDAO.AircraftGeometry.aspect_ratio — Method

aspect_ratio(wing :: AbstractWing)

Compute the aspect ratio of an AbstractWing.

AeroMDAO.AircraftGeometry.camber_CST — Function

camber_CST(α_c, α_t,
           (Δz_u, Δz_l) :: NTuple{2, Real},
           coeff_LE = 0.,
           n :: Integer = 40)

Define a cosine-spaced foil with $2n$ points using the Class Shape Transformation method on a Bernstein polynomial basis for the camber and thickness coordinates.

The foil is defined by arrays of coefficients $(α_c,~ α_t)$ for the upper and lower surfaces, trailing-edge spacing values $(Δz_u,~Δz_l)$, and a coefficient for the leading edge modifications.

AeroMDAO.AircraftGeometry.camber_coordinates — Method

camber_coordinates(coords :: Array{2, <: Real})

Generate the camber coordinates on the $x$-$z$ plane at $y = 0$.

AeroMDAO.AircraftGeometry.camber_coordinates — Method

camber_coordinates(wing :: HalfWing, n_s :: Integer, n_c :: Integer, flip = false)

Compute the camber coordinates of a HalfWing consisting of camber distributions of Foils and relevant geometric quantities, given numbers of spanwise $n_s$ and chordwise $n_c$ panels, with an option to flip the signs of the $y$-coordinates.

AeroMDAO.AircraftGeometry.camber_thickness — Function

camber_thickness(foil :: Foil, num :: Integer)

Compute the camber-thickness distribution of a Foil with cosine spacing.

AeroMDAO.AircraftGeometry.camber_thickness_to_coordinates — Method

camber_thickness_to_coordinates(xs, camber, thickness)

Convert the camber-thickness representation to 2-dimensional coordinates given the $x$-locations and their corresponding camber and thickness values.

AeroMDAO.AircraftGeometry.chord_coordinates — Method

chord_coordinates(wing :: AbstractWing, n_s :: Integer, n_c :: Integer, flip = false)

Compute the chord coordinates of a HalfWing consisting of Foils and relevant geometric quantities, given numbers of spanwise $n_s$ and chordwise $n_c$ panels, with an option to flip the signs of the $y$-coordinates.

AeroMDAO.AircraftGeometry.coordinates — Method

coordinates(foil :: Foil)

Generate the array of Foil coordinates.

AeroMDAO.AircraftGeometry.coordinates_to_camber_thickness — Function

coordinates_to_camber_thickness(coords, n = 40)

Convert 2-dimensional coordinates to its camber-thickness representation after cosine interpolation with $2n$ points.

AeroMDAO.AircraftGeometry.kulfan_CST — Function

kulfan_CST(alpha_u, alpha_l,
           (Δz_u, Δz_l) = (0., 0.),
           (LE_u, LE_l) = (0., 0.),
           n            = 40)

Define a cosine-spaced foil with $2n$ points using the Class Shape Transformation method on a Bernstein polynomial basis for the upper and lower coordinates.

The foil is defined by arrays of coefficients $(α_u,~ α_l)$ for the upper and lower surfaces, trailing-edge spacing values $(Δz_u,~ Δz_l)$, and coefficients for leading edge modifications.

AeroMDAO.AircraftGeometry.leading_edge — Function

leading_edge(wing :: HalfWing, flip = false)

Compute the leading edge coordinates of a HalfWing, with an option to flip the signs of the $y$-coordinates.

AeroMDAO.AircraftGeometry.mean_aerodynamic_center — Method

mean_aerodynamic_chord(wing :: AbstractWing)

Compute the coordinates of the mean aerodynamic center of an AbstractWing.

AeroMDAO.AircraftGeometry.mean_aerodynamic_chord — Method

mean_aerodynamic_chord(wing :: AbstractWing)

Compute the mean aerodynamic chord of an AbstractWing.

AeroMDAO.AircraftGeometry.mesh_cambers — Method

mesh_cambers(wing :: AbstractWing, n_s :: Integer, n_c :: Integer; spacings = symmetric_spacing(wing))

Mesh the camber distribution of a Wing into panels of $n_s$ spanwise divisions per section and $n_c$ chordwise divisions with an AbstractSpacing` distribution.

AeroMDAO.AircraftGeometry.mesh_horseshoes — Method

mesh_horseshoes(wing :: AbstractWing, n_s :: Vector{Integer}, n_c :: Integer; flip = false)

Mesh the span and chord distributions of an AbstractWing with $n_s$ spanwise divisions per section and $n_c$ chordwise divisions.

AeroMDAO.AircraftGeometry.mesh_wing — Method

mesh_horseshoes(wing :: AbstractWing, n_s :: Vector{Integer}, n_c :: Integer; flip = false)

Mesh the span and airfoil coordinate distributions of an AbstractWing with $n_s$ spanwise divisions per section and $n_c$ chordwise divisions.

AeroMDAO.AircraftGeometry.naca4 — Function

naca4(digits :: NTuple{4, <: Real}, n :: Integer; sharp_trailing_edge :: Bool)

Generate the coordinates of a NACA 4-digit series profile.

AeroMDAO.AircraftGeometry.projected_area — Method

projected_area(wing :: AbstractWing)

Compute the projected area of an AbstractWing` by summing the trapezoidal areas.

AeroMDAO.AircraftGeometry.read_foil — Method

read_foil(path :: String; header = true)

Read a '.dat' file consisting of 2D coordinates, for an airfoil as an array of SVectors, with an optional argument to skip the header.

AeroMDAO.AircraftGeometry.scale_foil — Method

scale_foil(foil :: Foil, scale)

Scale the coordinates of a Foil to a scaling value.

AeroMDAO.AircraftGeometry.span — Method

span(wing :: AbstractWing)

Compute the planform span of an AbstractWing.

AeroMDAO.AircraftGeometry.split_foil — Method

split_foil(foil :: Foil)

Split the Foil coordinates into upper and lower surfaces.

AeroMDAO.AircraftGeometry.surface_coordinates — Method

surface_coordinates(wing :: HalfWing, n_s :: Integer, n_c :: Integer, flip = false)

Compute the surface coordinates of a HalfWing consisting of Foils and relevant geometric quantities, given numbers of spanwise $n_s$ and chordwise $n_c$ panels, with an option to flip the signs of the $y$-coordinates.

AeroMDAO.AircraftGeometry.thickness_coordinates — Method

thickness_coordinates(coords :: Array{2, <: Real})

Generate the thickness coordinates on the $x$-$z$ plane at $y = 0$.

AeroMDAO.AircraftGeometry.wing_bounds — Function

wing_bounds(wing :: HalfWing, flip = false)

Compute the leading and trailing edge coordinates of a HalfWing, with an option to flip the signs of the $y$-coordinates.

AeroMDAO.AircraftGeometry.wing_bounds — Method

wing_bounds(wing :: Wing)

Return the leading and trailing edge coordinates of a Wing.

AeroMDAO.MathTools.cosine_spacing — Function

cosine_spacing(coordinates, n = 40)

Interpolate a foil profile's coordinates by projecting the x-coordinates of a circle onto the geometry with $2n$ points.

AeroMDAO.MathTools.cosine_spacing — Method

cosine_spacing(foil :: Foil, num :: Integer)

Return a Foil with cosine spacing for a given number of points.

AeroMDAO.properties — Method

properties(wing :: AbstractWing)

Compute the generic properties of interest (span, area, etc.) of an AbstractWing.

Panel Geometry API

AeroMDAO.PanelGeometry.Panel3D — Type

Panel3D(p1, p2, p3, p4)

Four Cartesian coordinates p1, p2, p3, p4 representing corners of a panel in 3 dimensions. The following commutative diagram (math joke) depicts the order:

z → y
↓
x
        p1 —→— p4
        |       |
        ↓       ↓
        |       |
        p2 —→— p3

AeroMDAO.PanelGeometry.make_panels — Method

make_panels(xyzs)

Convert an array of coordinates corresponding to a wing, ordered from root to tip and leading-edge to trailing-edge, into panels.

AeroMDAO.PanelGeometry.midpoint — Method

midpoint(panel :: Panel3D)

Compute the midpoint of a Panel3D.

AeroMDAO.PanelGeometry.panel_area — Method

panel_area(panel :: Panel3D)

Compute the (possibly non-planar, hence nonsensical) area of a Panel3D.

AeroMDAO.PanelGeometry.panel_coordinates — Method

panel_coordinates(panel :: Panel3D)

Compute the coordinates of a Panel3D.

AeroMDAO.PanelGeometry.panel_normal — Method

panel_normal(panel :: Panel3D)

Compute the normal vector of a Panel3D.

AeroMDAO.PanelGeometry.reflect_xz — Method

reflect_xz(panel :: Panel3D)

Reflect a Panel3D with respect to the $x$-$z$ plane of its reference coordinate system.

AeroMDAO.PanelGeometry.transform — Method

transform(panel :: Panel3D, rotation, translation)

Perform an affine transformation on the coordinates of a Panel3D given a rotation matrix and translation vector.

AeroMDAO.PanelGeometry.transform_normal — Method

transform_normal(panel :: Panel3D, h_l, g_l)

Transform the normal vector of a Panel3D about the hinge axis $ĥ_l$ by the control gain $g_l$.

AeroMDAO.PanelGeometry.wetted_area — Method

wetted_area(panels :: Array{Panel3D})

Compute the total wetted area by summing the areas of an array of Panel3D.