Structures API

Beams API

AeroFuse.Beams.Material — Method

Material(E, J, σ_max, ρ)
Material(; elastic_modulus, shear_modulus, yield_stress, density)

Define a Material with positive, real elastic modulus $E$, shear modulus $G$, yield stress $σ_{\max}$, and density $ρ$.

The default assignments for the named variables are set to the properties of aluminium?

Arguments

elastic_modulus :: Real = 85e9: Elastic modulus.

shear_modulus :: Real = 25e9: Shear modulus.

yield_stress :: Real = 350e6: Yield stress.

density :: Real = 1.6e3: Density.

AeroFuse.Beams.Tube — Type

Tube(material :: Material, length, radius, thickness)

Define a hollow tube of fixed radius with a given material, length, and thickness.

AeroFuse.Beams.tube_stiffness_matrix — Function

tube_stiffness_matrix(E, G, A, Iyy, Izz, J, L)
tube_stiffness_matrix(E, G, A, Iyy, Izz, J, L, num)
tube_stiffness_matrix(tube :: Tube)
tube_stiffness_matrix(tube :: Vector{Tube})
tube_stiffness_matrix(x :: Matrix{Real})

Generate the stiffness matrix using the properties of a tube.

The required properties are the elastic modulus $E$, shear modulus $G$, area $A$, moments of inertia about the $y-$ and $z-$ axes $I_{yy}, I_{zz}$, polar moment of inertia $J$, length $L$. A composite stiffness matrix is generated with a specified number of elements.