Aerodynamics API

Doublet-Source Panel API

AeroFuse.DoubletSource.boundary_vector — Method

boundary_vector(panels, u)

Create the vector for the boundary condition of the problem given an array of Panel2Ds and velocity $u$.

AeroFuse.DoubletSource.doublet_matrix — Method

doublet_matrix(panels_1, panels_2)

Create the matrix of doublet potential influence coefficients between pairs of panels₁ and panels₂.

AeroFuse.DoubletSource.influence_matrix — Method

influence_matrix(panels, wake_panel :: AbstractPanel2D)

Assemble the Aerodynamic Influence Coefficient matrix consisting of the doublet matrix, wake vector, Kutta condition given Panel2Ds and the wake panel.

AeroFuse.DoubletSource.kutta_condition — Method

kutta_condition(panels)

Create the vector describing Morino's Kutta condition given Panel2Ds.

AeroFuse.DoubletSource.solve_linear — Method

solve_linear(panels, u, wakes)

Solve the linear aerodynamic system given the array of Panel2Ds, a velocity $\vec U$, a vector of wake Panel2Ds, and an optional named bound for the length of the wake.

The system of equations $A[φ] = [\vec{U} ⋅ n̂] - B[σ]$ is solved, where $A$ is the doublet influence coefficient matrix, $φ$ is the vector of doublet strengths, $B$ is the source influence coefficient matrix, and $σ$ is the vector of source strengths.

AeroFuse.DoubletSource.solve_linear — Method

solve_linear(panels, u, sources, bound)

Solve the system of equations $[AIC][φ] = [\vec{U} ⋅ n̂] - B[σ]$ condition given the array of Panel2Ds, a velocity $\vec U$, a condition whether to disable source terms ($σ = 0$), and an optional named bound for the length of the wake.

AeroFuse.DoubletSource.source_matrix — Method

doublet_matrix(panels_1, panels_2)

Create the matrix of source potential influence coefficients between pairs of panels₁ and panels₂.

AeroFuse.DoubletSource.source_strengths — Method

source_strengths(panels, freestream)

Create the vector of source strengths for the Dirichlet boundary condition $σ = \vec U_{\infty} ⋅ n̂$ given Panel2Ds and a Uniform2D.

AeroFuse.DoubletSource.wake_vector — Method

wake_vector(woke_panel :: AbstractPanel2D, panels)

Create the vector of doublet potential influence coefficients from the wake on the panels given the wake panel and the array of Panel2Ds.

AeroFuse.surface_velocities — Method

surface_velocities(panels, φs, u, sources :: Bool)

Compute the surface speeds and panel distances given the array of Panel2Ds, their associated doublet strengths $φ$s, the velocity $u$, and a condition whether to disable source terms ($σ = 0$).

Vortex Lattice API

AeroFuse.VortexLattice.Horseshoe — Type

Horseshoe(r1, r2, rc, normal, chord)

Define a horseshoe vortex with a start and endpoints $r₁, r₂$ for the bound leg, a collocation point $r$, a normal vector $n̂$, and a finite core size.

The finite core setup is not implemented for now.

AeroFuse.VortexLattice.References — Type

References(V, ρ, μ, S, b, c, r)
References(; speed, density, viscosity,
             sound_speed, area, span, 
             chord, location)

Define reference values with speed $V$, density $ρ$, dynamic viscosity $μ$, area $S$, span $b$, chord $c$, location $r$ for a vortex lattice analysis.

Arguments

speed :: Real = 1.: Speed (m/s)

density :: Real = 1.225: Density (m)

viscosity :: Real = 1.5e-5: Dynamic viscosity (m)

sound_speed :: Real = 330.: Speed of sound (m/s)

area :: Real = 1.: Area (m²)

span :: Real = 1.: Span length (m)

chord :: Real = 1.: Chord length (m)

location :: Vector{Real} = [0,0,0]: Position

AeroFuse.VortexLattice.VortexLatticeSystem — Type

VortexLatticeSystem(
    aircraft, 
    fs :: Freestream, 
    refs :: References, 
    compressible = false, 
    axes = Geometry()
)

Construct a VortexLatticeSystem for analyzing inviscid aerodynamics of an aircraft (must be a ComponentArray of Horseshoes or VortexRings) with Freestream conditions and References for non-dimensionalization. Options are provided for compressibility corrections via the Prandtl-Glauert transformation (false by default) and axis system for computing velocities and forces (Geometry by default).

AeroFuse.VortexLattice.VortexLatticeSystem — Type

VortexLatticeSystem

A system consisting of the relevant variables for a vortex lattice analysis for post-processing.

Arguments

The accessible fields are:

vortices: The array of vortices (currently Horseshoe).

circulations: The circulation strengths of the vortices obtained by solving the linear system.

influence_matrix: The influence matrix of the linear system.

boundary_vector: The boundary condition corresponding to the right-hand-side of the linear system.

freestream :: Freestream: The freestream conditions.

reference :: References: The reference values.

AeroFuse.VortexLattice.VortexRing — Type

VortexRing(r1, r2, r3, r4, r_c, n̂, ε)

A vortex ring consisting of four points $r_i, i = 1,…,4$, a collocation point $r_c$, a normal vector $n̂$, and a core size $ε$. The following convention is adopted:

    r1 —front leg→ r4
    |               |
left leg       right leg
    ↓               ↓
    r2 —back leg-→ r3

AeroFuse.VortexLattice.bound_leg_center — Method

bound_leg_center(hs :: Horseshoe)

Compute the midpoint of the bound leg of a Horseshoe.

AeroFuse.VortexLattice.bound_leg_vector — Method

bound_leg_vector(hs :: Horseshoe)

Compute the direction vector of the bound leg of a Horseshoe.

AeroFuse.VortexLattice.bound_velocity — Method

bound_velocity(r, hs :: Horseshoes, Γ, u_hat)

Compute the induced velocity at a point $r$ from the bound leg with constant strength $Γ$ of a given Horseshoe hs.

AeroFuse.VortexLattice.boundary_condition — Method

boundary_condition(horseshoes, U, Ω)

Assemble the boundary condition vector given an array of Horseshoes, the freestream velocity $U$, and a quasi-steady rotation vector $Ω$.

AeroFuse.VortexLattice.compute_farfield_forces — Method

compute_farfield_forces(Δφs, Δs, ∂φ_∂n, θs, V, ρ)

Compute the aerodynamic forces in the Trefftz plane given cumulative doublet strengths $Δφ$s, Trefftz panel lengths $Δs$, doublet-normal directional derivatives $∂φ/∂n$, Trefftz panel angles $θ$s, the freestream speed $V$ and density $ρ$.

AeroFuse.VortexLattice.doublet_normal_derivatives — Method

doublet_normal_derivatives(wake_points, Δφs, normals)

Compute the normal derivative strengths of the doublets given the wake points, net doublet strengths $Δφ$s, and associated normal vectors.

AeroFuse.VortexLattice.farfield — Method

farfield(system :: VortexLatticeSystem)

Compute the total farfield force coefficients for all components of the VortexLatticeSystem. These are in wind axes by definition.

AeroFuse.VortexLattice.farfield_coefficients — Method

farfield_coefficients(system :: VortexLatticeSystem)

Compute the total farfield force coefficients for all components of the VortexLatticeSystem. These are in wind axes by definition.

AeroFuse.VortexLattice.farfield_forces — Method

farfield_forces(Γs, horseshoes, freestream, ρ)

Compute the aerodynamic forces in the Trefftz plane normal to the freestream given horseshoes, their associated strengths $Γ$s, and a density $ρ$.

AeroFuse.VortexLattice.farfield_forces — Method

farfield_forces(system :: VortexLatticeSystem)

Compute the farfield forces in wind axes for all components of the VortexLatticeSystem.

AeroFuse.VortexLattice.farfield_influence_matrix — Method

farfield_influence_matrix(centers, normals, points)

Compute the aerodynamic influence coefficient matrix of the wake in the Trefftz plane given the center points, normal vectors, and the points of the wake panels.

AeroFuse.VortexLattice.farfield_velocity — Method

farfield_velocity(r_i, r_j, Γ_j)

Compute the induced velocity at point $r_i$ of the wake in the Trefftz plane due to the strength $Γ_j$ at $r_j$.

AeroFuse.VortexLattice.flip_xz — Method

flip_xz(vector)

Reflect the $x$- and $z$- coordinates of a given 3-dimensional vector about the $y$-$z$ and $x$-$y$ planes respectively for the representation in body axes.

AeroFuse.VortexLattice.freestream_derivatives — Method

freestream_derivatives(
    system :: VortexLatticeSystem,
    name = :aircraft,
    axes = Wind(),
    print = false,
    print_components = false
)

Obtain the force and moment coefficients of the components of a VortexLatticeSystem and their derivatives with respect to freestream values: Mach $M$ (if compressible), angles of attack $α$ and sideslip $β$, and non-dimensionalized angular velocity rates $p̄, q̄, r̄$.

The axes can be changed by passing any axis system (such as Body(), Geometry(), Wind(), Stability()) to the named axes argument. The force and moment coefficients are reported in wind axes by default.

AeroFuse.VortexLattice.geometry_to_stability_axes — Method

geometry_to_stability_axes(coords, α)

Convert coordinates from geometry to stability axes with angle $α$.

AeroFuse.VortexLattice.geometry_to_wind_axes — Method

geometry_to_wind_axes(coords, α, β)
geometry_to_wind_axes(vor :: AbstractVortex, α, β)

Convert coordinates from geometry axes to wind axes for given angles of attack $α$ and sideslip $\beta.$

AeroFuse.VortexLattice.influence_matrix — Method

influence_matrix(horseshoes)
influence_matrix(horseshoes, u_hat)

Assemble the Aerodynamic Influence Coefficient (AIC) matrix given an array of Horseshoes and the freestream direction $̂u$.

AeroFuse.VortexLattice.nearfield — Method

nearfield(system :: VortexLatticeSystem)

Compute the total force and moment coefficients in wind axes for all components of the VortexLatticeSystem.

AeroFuse.VortexLattice.nearfield_coefficients — Method

nearfield_coefficients(system :: VortexLatticeSystem)

Compute the force and moment coefficients in wind axes for all components of the VortexLatticeSystem.

AeroFuse.VortexLattice.print_coefficients — Function

print_coefficients(system :: VortexLatticeSystem, name = :aircraft;
                   components = false)

Print a pretty table of the total nearfield and farfield coefficients of a VortexLatticeSystem with an optional name.

A named Boolean argument components is provided to also enable the printing of any possible components.

AeroFuse.VortexLattice.print_coefficients — Function

print_coefficients(nf_coeffs, ff_coeffs, name = "")

Print a pretty table of the nearfield and farfield coefficients with an optional name.

AeroFuse.VortexLattice.print_derivatives — Function

print_derivatives(nf_coeffs, ff_coeffs, name = "";
                  farfield = false,
                  axes = "")

Print a pretty table of the aerodynamic coefficients and derivatives with an optional name and named arguments for enabling printing of farfield coefficients and the axis name.

AeroFuse.VortexLattice.project_yz — Method

project_yz(vector)

Project a given 3-dimensional vector or into the $y$-$z$ plane.

AeroFuse.VortexLattice.reflect_xz — Method

reflect_xz(vector)

Reflect the $y$-coordinate of a given 3-dimensional vector about the $x$-$z$ plane.

AeroFuse.VortexLattice.stability_to_geometry_axes — Method

geometry_to_stability_axes(coords, α)

Convert coordinates from stability to geometry axes with angle $α$.

AeroFuse.VortexLattice.stream_velocity — Method

stream_velocity(r, horseshoes, Γs, V, Ω)

Evaluate the total induced velocity at a point r given Horseshoes, vortex strengths Γs, rotation rates Ω, and freestream flow vector freestream in the global reference frame.

AeroFuse.VortexLattice.streamlines — Method

streamlines(point, freestream :: Freestream, horseshoes, Γs, length, num_steps)

Compute the streamlines from a given starting point, a Freestream, Horseshoes and their associated strengths Γs with a specified length of the streamline and number of evaluation points.

AeroFuse.VortexLattice.surface_dynamics — Method

surface_dynamics(system :: VortexLatticeSystem; 
                 axes   :: AbstractAxisSystem = Geometry())

Compute the forces and moments for all components of the VortexLatticeSystem in the reference axis system.

AeroFuse.VortexLattice.surface_forces — Method

surface_forces(hs_comp, Γ_comp, horseshoes, Γs, U, Ω, ρ)
surface_forces(horseshoes, Γs, U, Ω, ρ)

Compute the surface forces via the local Kutta-Jowkowsky theorem.

For a given array of vortex type hs_comp and their associated vortex strengths $Γ_c$ for which to compute the forces, the arrays of horseshoes and vortex strengths $Γ$s of the entire aircraft, the freestream flow vector $U$, rotation rates $Ω$, and a density $ρ$, the velocities are evaluated at the midpoint of the bound leg of each horseshoe, excluding the contribution of the bound leg vortex.

The second variant simply sets hs_comp = horseshoes and Γ_comp = Γs.

AeroFuse.VortexLattice.surface_forces — Method

surface_forces(system :: VortexLatticeSystem; 
               axes   :: AbstractAxisSystem = Geometry())

Compute the forces for all components of the VortexLatticeSystem in the reference axis system.

AeroFuse.VortexLattice.surface_moments — Method

surface_moments(system :: VortexLatticeSystem; 
                axes   :: AbstractAxisSystem = Geometry())

Compute the moments for all components of the VortexLatticeSystem in the reference axis system.

AeroFuse.VortexLattice.trailing_velocity — Method

trailing_velocity(r, hs :: Horseshoes, Γ, u_hat)

Compute the induced velocity at a point $r$ from the semi-infinite trailing legs with constant strength $Γ$ of a given Horseshoe hs.

AeroFuse.VortexLattice.transform — Method

transform(hs :: Horseshoe, T :: LinearMap)

Generate a new Horseshoe with the points and normal vectors transformed by the LinearMap $T$.

AeroFuse.VortexLattice.transform — Method

transform(ring :: VortexRing, T :: LinearMap)

Generate a new VortexRing with the points and normal vectors transformed by the LinearMap $T$.

AeroFuse.VortexLattice.trefftz_plane_quantities — Method

trefftz_plane_quantities(vortices, α, β)

Project Horseshoes into the Trefftz plane aligned with the wind axes angles $α,~β$, and compute normal vectors, projection angles and lengths.

AeroFuse.solve_linear — Method

solve_linear(horseshoes, normals, U, Ω)

Evaluate and return the vortex strengths $Γ$s given an array of Horseshoes, their associated normal vectors, the velocity vector $U$, and the quasi-steady rotation vector $Ω$.

AeroFuse.surface_coefficients — Method

surface_coefficients(system :: VortexLatticeSystem; 
                     axes   :: AbstractAxisSystem = Wind())

Compute the force and moment coefficients on the surface given the VortexLatticeSystem in the reference axis system.

AeroFuse.surface_velocities — Method

surface_velocities(system :: VortexLatticeSystem; 
                   axes   :: AbstractAxisSystem = Geometry())

Compute the induced velocities for all components of the VortexLatticeSystem in the reference axis system.

AeroFuse.velocity — Function

velocity(r, hs :: Horseshoe, Γ, u_hat = [1.,0.,0.])

Compute the induced velocity at a point $r$ of a given Horseshoe hs with a bound leg of constant strength $Γ$ and semi-infinite trailing legs pointing in a given direction $û$, by default û = x̂.

AeroFuse.velocity — Function

velocity(r, ring :: VortexRing, Γ)

Computes the velocity at a point $r$ induced by a VortexRing with constant strength $Γ$.

AeroFuse.velocity — Method

velocity(freestream :: Freestream, ::Body)

Compute the velocity of Freestream in the body reference frame.