Aerodynamics API

Doublet-Source Panel API

AeroMDAO.DoubletSource.boundary_vector — Method

boundary_vector(panels, u)

Create the vector for the boundary condition of the problem given an array of Panel2Ds and velocity $u$.

AeroMDAO.DoubletSource.doublet_matrix — Method

doublet_matrix(panels_1, panels_2)

Create the matrix of doublet potential influence coefficients between pairs of panels1 and panels2.

AeroMDAO.DoubletSource.influence_matrix — Method

influence_matrix(panels, wake_panel)

Assemble the Aerodynamic Influence Coefficient matrix consisting of the doublet matrix, wake vector, Kutta condition given Panel2Ds and the wake panel.

AeroMDAO.DoubletSource.kutta_condition — Method

kutta_condition(panels)

Create the vector describing Morino's Kutta condition given Panel2Ds.

AeroMDAO.DoubletSource.solve_linear — Method

solve_linear(panels, u, sources, bound)

Solve the system of equations $[AIC][\phi] = [\vec{U} \cdot \hat{n}] - B[\sigma]$ condition given the array of Panel2Ds, a velocity $\vec U$, a condition whether to disable source terms ($σ = 0$), and an optional named bound for the length of the wake.

AeroMDAO.DoubletSource.solve_linear — Method

solve_linear(panels, u, wakes, bound = 1e5)

Solve the linear aerodynamic system given the array of Panel2Ds, a velocity $\vec U$, a vector of wake Panel2Ds, and an optional named bound for the length of the wake.

The system of equations $A[\phi] = [\vec{U} \cdot \hat{n}] - B[\sigma]$ is solved, where $A$ is the doublet influence coefficient matrix, $\phi$ is the vector of doublet strengths, $B$ is the source influence coefficient matrix, and $\sigma$ is the vector of source strengths.

AeroMDAO.DoubletSource.source_matrix — Method

source_matrix(panels_1, panels_2)

Create the matrix of source potential influence coefficients between pairs of panels_1 and panels_2.

AeroMDAO.DoubletSource.source_strengths — Method

source_strengths(panels, freestream)

Create the vector of source strengths for the Dirichlet boundary condition $σ = \vec U_{\infty} \cdot \hat{n}$ given Panel2Ds and a Uniform2D.

AeroMDAO.DoubletSource.wake_vector — Method

wake_vector(wake_panel, panels)

Create the vector of doublet potential influence coefficients from the wake on the panels given the wake panel and the array of Panel2Ds.

AeroMDAO.surface_velocities — Method

surface_velocities(panels, φs, u, sources :: Bool)

Compute the surface speeds and panel distances given the array of Panel2Ds, their associated doublet strengths $φ$s, the velocity $u$, and a condition whether to disable source terms ($σ = 0$).

Vortex Lattice API

AeroMDAO.VortexLattice.Freestream — Type

Freestream(α, β, Ω)
Freestream(α, β, Ω_x, Ω_y, Ω_z)
Freestream(U, Ω)
Freestream(; 
             alpha = 0., 
             beta  = 0., 
             omega = [0,0,0]
           )

Define freestream conditions with angle of attack $α$ (degrees), sideslip angle $β$ (degrees), and (quasi-steady) rotation vector $Ω$ for a vortex lattice analysis.

Alternatively, provide the velocity vector $U$, which is normalized to determine the angles.

AeroMDAO.VortexLattice.Horseshoe — Type

Horseshoe(bound_leg, collocation_point, normal, chord)

Define a horseshoe vortex based on a bound leg, a collocation point, a normal vector, and a "chord length".

AeroMDAO.VortexLattice.Horseshoe — Type

Horseshoe(panel :: Panel3D, normal, drift = zeros(3))

Generate a Horseshoe corresponding to a Panel3D, an associated normal vector, and a "drift velocity".

AeroMDAO.VortexLattice.Line — Type

Line(r1 :: SVector{3,<: Real}, r2 :: SVector{3,<: Real})

Define a line segment based on two 3-dimensional vectors.

AeroMDAO.VortexLattice.References — Type

References(V, ρ, μ, S, b, c, r)
References(; speed, density, viscosity,
             sound_speed, area, span, 
             chord, location)

Define reference values with speed $V$, density $ρ$, dynamic viscosity $μ$, area $S$, span $b$, chord $c$, location $r$ for a vortex lattice analysis.

Arguments

speed :: Real = 1.: Speed (m/s)

density :: Real = 1.225: Density (m)

viscosity :: Real = 1.5e-5: Dynamic viscosity (m)

sound_speed :: Real = 330.: Speed of sound (m/s)

span :: Real = 1.: Span length (m)

area :: Real = 1.: Area (m²)

chord :: Real = 1.: Chord length (m)

location :: Vector{Real} = [0,0,0]: Position

AeroMDAO.VortexLattice.VortexRing — Type

VortexRing{T <: Real}(left_leg, bound_leg, back_leg, right_leg)

A vortex ring type consisting of four finite vortex line filaments connected in a loop.

AeroMDAO.VortexLattice.VortexRing — Method

Constructor for vortex rings on a Panel3D using Lines. The following convention is adopted:

    p1 —bound_leg→ p4
    |               |
left_leg       right_leg
    ↓               ↓
    p2 —back_leg→ p3

AeroMDAO.VortexLattice.body_frame_velocity — Method

body_frame_velocity(freestream :: Freestream)

Compute the velocity of Freestream in the body reference frame.

AeroMDAO.VortexLattice.bound_leg — Method

bound_leg(horseshoe :: Horseshoe)

Getter for bound leg field of a Horseshoe.

AeroMDAO.VortexLattice.bound_leg — Method

bound_leg(panel :: Panel3D)

Compute the bound leg for a Panel3D, for horseshoes/vortex rings.

AeroMDAO.VortexLattice.bound_leg_center — Method

Compute the midpoint of the bound leg of a Horseshoe.

AeroMDAO.VortexLattice.bound_leg_vector — Method

Compute the direction vector of the bound leg of a Horseshoe.

AeroMDAO.VortexLattice.boundary_condition — Method

boundary_condition(velocities, normals)

Compute the projection of a velocity vector with respect to normal vectors of panels. Corresponds to construction of the boundary condition for the RHS of the AIC system.

AeroMDAO.VortexLattice.compute_farfield_forces — Method

compute_farfield_forces(Δφs, Δs, ∂φ_∂n, θs, V, ρ)

Compute the aerodynamic forces in the Trefftz plane given cumulative doublet strengths $Δφ$s, Trefftz panel lengths $Δs$, doublet-normal directional derivatives $∂φ/∂n$, Trefftz panel angles $θ$s, the freestream speed $V$ and density $ρ$.

AeroMDAO.VortexLattice.doublet_normal_derivatives — Method

doublet_normal_derivatives(wake_lines :: Vector{<: Line}, Δφs, normals)

Compute the normal derivative strengths of the doublets given the wake Lines, net doublet strengths $Δφ$s, and associated normal vectors.

AeroMDAO.VortexLattice.farfield_forces — Method

farfield_forces(Γs, horseshoes, freestream, ρ)

Compute the aerodynamic forces in the Trefftz plane normal to the freestream given horseshoes, their associated strengths $Γ$s, and a density $ρ$.

AeroMDAO.VortexLattice.farfield_influence_matrix — Method

farfield_influence_matrix(centers, normals, points)

Compute the aerodynamic influence coefficient matrix of the wake in the Trefftz plane given the center points, normal vectors, and the points of the wake panels.

AeroMDAO.VortexLattice.farfield_velocity — Method

farfield_velocity(r_i, r_j, Γ_j)

Compute the induced velocity at point $r_i$ of the wake in the Trefftz plane due to the strength $Γ_j$ at $r_j$.

AeroMDAO.VortexLattice.flip_xz — Method

flip_xz(vector)

Reflect the $x$- and $z$- coordinates of a given 3-dimensional vector about the $y$-$z$ and $x$-$y$ planes respectively for the representation in body axes.

AeroMDAO.VortexLattice.geometry_to_stability_axes — Method

geometry_to_stability_axes(coords, α)

Convert coordinates from geometry to stability axes with angle $α$.

AeroMDAO.VortexLattice.geometry_to_wind_axes — Method

geometry_to_wind_axes(coords, α, β)
geometry_to_wind_axes(line :: Line, α, β)

Convert coordinates from geometry axes to wind axes for given angles of attack $α$ and sideslip $\beta.$

AeroMDAO.VortexLattice.horseshoe_point — Method

collocation_point(panel :: Panel3D)

Compute the collocation point of a Panel3D for horseshoes/vortex rings, which is the 3-quarter point on each side in the $x$-$z$ plane.

AeroMDAO.VortexLattice.influence_coefficient — Function

influence_coefficient(r, horseshoe, normal, V_hat, symmetry)

Compute the influence coefficient of the velocity of a Horseshoe with trailing lines in a given direction $û$ at a point $r$ projected to a normal vector.

AeroMDAO.VortexLattice.influence_matrix — Function

influence_matrix(horseshoes, collocation_points, normals, V_hat)

Assemble the Aerodynamic Influence Coefficient (AIC) matrix given horseshoes, collocation points, associated normal vectors, and a unit vector representing the freestream.

AeroMDAO.VortexLattice.project_yz — Method

project_yz(vector)
project_yz(line :: Line)

Project a given 3-dimensional vector or Line into the $y$-$z$ plane.

AeroMDAO.VortexLattice.reflect_xz — Method

reflect_xz(vector)
reflect_xz(line :: Line)
reflect_xz(horseshoe :: Horseshoe)

Reflect the $y$-coordinate of a given 3-dimensional vector (or $Line$ or $Horseshoe$) about the $x$-$z$ plane.

AeroMDAO.VortexLattice.stability_to_geometry_axes — Method

geometry_to_stability_axes(coords, α)

Convert coordinates from stability to geometry axes with angle $α$.

AeroMDAO.VortexLattice.stream_velocity — Method

stream_velocity(r, horseshoes, Γs, V, Ω)

Evaluate the total induced velocity at a point r given Horseshoes, vortex strengths Γs, rotation rates Ω, and freestream flow vector freestream in the global reference frame.

AeroMDAO.VortexLattice.streamlines — Method

streamlines(point, freestream :: Freestream, horseshoes, Γs, length, num_steps)

Compute the streamlines from a given starting point, a Freestream, Horseshoes and their associated strengths Γs with a specified length of the streamline and number of evaluation points.

AeroMDAO.VortexLattice.sum_vortices — Method

sum_vortices(r, vortex_lines :: Array{Line}, Γ)

Sums the velocities evaluated at a point $r$ of Lines with constant strength $Γ$.

AeroMDAO.VortexLattice.surface_dynamics — Method

surface_dynamics(system :: VortexLatticeSystem; 
                 axes   :: AbstractAxisSystem = Geometry())

Compute the forces and moments on the surface given the VortexLatticeSystem after performing an analysis, and the reference axis system.

AeroMDAO.VortexLattice.surface_forces — Method

surface_forces(hs_comp, Γ_comp, horseshoes, Γs, U, Ω, ρ)

Compute the surface forces via the local Kutta-Jowkowsky theorem.

For a given array of vortex type hs_comp and their associated vortex strengths $Γ_c$ for which to compute the forces, the arrays of horseshoes and vortex strengths $Γ$s of the entire aircraft, the freestream flow vector $U$, rotation rates $Ω$, and a density $ρ$, the velocities are evaluated at the midpoint of the bound leg of each horseshoe, excluding the contribution of the bound leg.

AeroMDAO.VortexLattice.surface_forces — Method

surface_forces(system :: VortexLatticeSystem; 
               axes   :: AbstractAxisSystem = Geometry())

Compute the forces on the surface given the VortexLatticeSystem after performing an analysis, and the reference axis system.

AeroMDAO.VortexLattice.surface_moments — Method

surface_moments(system :: VortexLatticeSystem; 
                axes   :: AbstractAxisSystem = Geometry())

Compute the moments on the surface given the VortexLatticeSystem after performing an analysis, and the reference axis system.

AeroMDAO.VortexLattice.trailing_velocity — Method

trailing_velocity(r, horseshoe, Γ, V_hat)

Compute the induced velocities at a point $r$ of a given Horseshoe with constant strength $Γ$ and trailing legs pointing in a given direction $û$, excluding the bound leg contribution.

AeroMDAO.VortexLattice.trefftz_plane_quantities — Method

trefftz_plane_quantities(horseshoes :: AbstractArray{<: Horseshoe}, α, β)

Project Horseshoes into the Trefftz plane aligned with the wind axes angles $α,~β$, and compute normal vectors, projection angles and lengths.

AeroMDAO.VortexLattice.vortex_ring — Method

Helper function to compute the vortex ring given four points following Panel3D ordering.

AeroMDAO.VortexLattice.vortex_ring — Method

Computes the vortex rings on a Panel3D.

AeroMDAO.VortexLattice.wind_to_geometry_axes — Method

wind_to_geometry_axes(coords, α, β)

Convert coordinates from wind axes to geometry axes for given angles of attack $α$ and sideslip \beta.``

AeroMDAO.solve_linear — Function

solve_linear(horseshoes, normals, U, Ω)

Evaluate and return the vortex strengths $Γ$s given Horseshoes, their associated normal vectors (not necessarily the same as the panels' normals), the speed $U$ and rotation vector $Ω$.

AeroMDAO.surface_velocities — Method

surface_velocities(system :: VortexLatticeSystem; 
                   axes   :: AbstractAxisSystem = Geometry())

Compute the velocities on the surface given the VortexLatticeSystem after performing an analysis, and the reference axis system.

AeroMDAO.velocity — Function

velocity(r, horseshoe, Γ, V_hat)

Compute the induced velocities at a point $r$ of a given Horseshoe with constant strength $Γ$ and trailing legs pointing in a given direction $û$.

AeroMDAO.velocity — Method

Computes the induced velocities at a point $r$ of a VortexRing with constant strength $Γ$.

AeroMDAO.velocity — Method

velocity(freestream :: Freestream)

Compute the velocity of a Freestream.