Geometry API

Aircraft Geometry API

AeroFuse.AircraftGeometry.AbstractSpacing — Type

An abstract type to define custom spacing distributions.

AeroFuse.AircraftGeometry.Foil — Type

Foil(x, y, name = "")
Foil(coordinates, name = "")

Structure consisting of foil coordinates in 2 dimensions with an optional name.

The coordinates should be provided in counter-clockwise format, viz. from the trailing edge of the upper surface to the trailing edge of the lower surface.

AeroFuse.AircraftGeometry.Wing — Type

Wing(
    foils :: Vector{Foil}, 
    chords, 
    twists, 
    spans, 
    dihedrals, 
    sweeps,
    position = zeros(3),
    angle    = 0.
    axis     = [0.,1.,0.]
)

Definition for a Wing consisting of $N+1$ Foils, their associated chord lengths $c$ and twist angles $ι$, for $N$ sections with span lengths $b$, dihedrals $δ$ and leading-edge sweep angles $Λ_{LE}$, with all angles in degrees.

AeroFuse.AircraftGeometry.WingMesh — Type

WingMesh(
    surface :: AbstractWing, 
    n_span :: Vector{Integer}, n_chord :: Integer;
    span_spacing :: AbstractSpacing = symmetric_spacing(surface)
)

Define a container to generate meshes and panels for a given AbstractWing with a specified distribution of number of spanwise panels, and a number of chordwise panels.

Optionally a combination of AbstractSpacing types (Sine(), Cosine(), Uniform()) can be provided to the named argument span_spacing, either as a singleton or as a vector with length equal to the number of spanwise sections. By default, the combination is [Sine(), Cosine(), ..., Cosine()].

For surface coordinates, the wing mesh will have (nchord - 1) * 2 chordwise panels from TE-LE-TE and (nspan * 2) spanwise panels.

AeroFuse.AircraftGeometry.WingSection — Method

WingSection(; span, dihedral, sweep, taper, root_chord,
              root_twist, tip_twist, root_foil, tip_foil,
              position, angle, axis)

Define a Wing in the $x$-$z$ plane, with optional Boolean arguments for symmetry and flipping in the plane.

Arguments

span :: Real = 1.: Span length

dihedral :: Real = 1.: Dihedral angle (degrees)

LE_sweep :: Real = 0.: Leading-edge sweep angle (degrees)

taper :: Real = 1.: Taper ratio of tip to root chord

root_chord :: Real = 1.: Root chord length

root_twist :: Real = 0.: Twist angle at root (degrees)

tip_twist :: Real = 0.: Twist angle at tip (degrees)

root_foil :: Foil = naca4((0,0,1,2)): Foil coordinates at root

tip_foil :: Foil = root_foil: Foil coordinates at tip

root_control :: NTuple{2} = (0., 0.75): (Angle, chord-length ratio) for adding a control surface at the root.

tip_control :: NTuple{2} = root_control: (Angle, chord-length ratio) for adding a control surface at the tip. Defaults to root control's settings.

position :: Vector{Real} = zeros(3): Position

angle :: Real = 0.: Angle of rotation (degrees)

axis :: Vector{Real} = [0.,1.,0.]: Axis of rotation

AeroFuse.AircraftGeometry.arc_length — Method

arc_length(foil :: Foil)

Compute the arc-length of a Foil.

AeroFuse.AircraftGeometry.aspect_ratio — Method

aspect_ratio(wing :: AbstractWing)

Compute the aspect ratio of an AbstractWing.

AeroFuse.AircraftGeometry.camber_CST — Function

camber_CST(α_c, α_t,
           (Δz_u, Δz_l) :: NTuple{2, Real},
           coeff_LE = 0.,
           n :: Integer = 40)

Define a cosine-spaced foil with $2n$ points using the Class Shape Transformation method on a Bernstein polynomial basis for the camber and thickness coordinates.

The foil is defined by arrays of coefficients $(α_c,~ α_t)$ for the upper and lower surfaces, trailing-edge spacing values $(Δz_u,~Δz_l)$, and a coefficient for the leading edge modifications at the nose.

AeroFuse.AircraftGeometry.camber_coordinates — Function

camber_coordinates(wing :: WingMesh, n_span = wing.num_span, n_chord = wing.num_chord)

Generate the camber coordinates of a WingMesh with default spanwise $n_s$ and chordwise $n_c$ panel distributions from the mesh.

AeroFuse.AircraftGeometry.camber_coordinates — Method

camber_coordinates(coords :: Array{2, <: Real})

Generate the camber coordinates on the $x$-$z$ plane at $y = 0$.

AeroFuse.AircraftGeometry.camber_panels — Method

camber_panels(wing_mesh :: WingMesh)

Generate the camber mesh as a matrix of Panel3D from a WingMesh.

AeroFuse.AircraftGeometry.camber_thickness — Function

camber_thickness(foil :: Foil, num :: Integer)

Compute the camber-thickness distribution of a Foil with cosine spacing.

AeroFuse.AircraftGeometry.camber_thickness_to_CST — Method

camber_thickness_to_CST(coords, num_dvs)

Convert camber-thickness coordinates to a specified number of CST variables by performing a least-squares solution.

AeroFuse.AircraftGeometry.camber_thickness_to_coordinates — Method

camber_thickness_to_coordinates(xs, camber, thickness)

Convert the camber-thickness representation to 2-dimensional coordinates given the $x$-locations and their corresponding camber and thickness values.

AeroFuse.AircraftGeometry.chord_coordinates — Function

chord_coordinates(wing :: WingMesh, n_span = wing.num_span, n_chord = wing.num_chord)

Generate the chord coordinates of a WingMesh with default spanwise $n_s$ and chordwise $n_c$ panel distributions from the mesh.

AeroFuse.AircraftGeometry.chord_panels — Method

chord_panels(wing_mesh :: WingMesh)

Generate the chord mesh as a matrix of Panel3D from a WingMesh.

AeroFuse.AircraftGeometry.control_surface — Method

control_surface(foil :: Foil; angle, hinge)

Modify a Foil to mimic a control surface by specifying a deflection angle (in degrees, clockwise-positive convention) and a normalized hinge $x$-coordinate $∈ [0,1]$ in terms of the chord length.

AeroFuse.AircraftGeometry.coordinates — Method

coordinates(foil :: Foil)

Generate the array of Foil coordinates.

AeroFuse.AircraftGeometry.coordinates_to_CST — Method

coordinates_to_CST(coords, num_dvs)

Convert coordinates to a specified number of CST variables by performing a least-squares solution.

AeroFuse.AircraftGeometry.coordinates_to_camber_thickness — Function

coordinates_to_camber_thickness(coords, n = 40)

Convert 2-dimensional coordinates to its camber-thickness representation after cosine interpolation with $2n$ points.

AeroFuse.AircraftGeometry.cosine_interpolation — Function

cosine_interpolation(foil :: Foil, n :: Integer = 40)

Interpolate a Foil profile's coordinates to a cosine by projecting the x-coordinates of a circle onto the geometry with $2n$ points.

AeroFuse.AircraftGeometry.kulfan_CST — Function

kulfan_CST(alpha_u, alpha_l,
           (Δz_u, Δz_l) = (0., 0.),
           (LE_u, LE_l) = (0., 0.),
           n            = 40)

Define a cosine-spaced foil with $2n$ points using the Class Shape Transformation method on a Bernstein polynomial basis for the upper and lower coordinates.

The foil is defined by arrays of coefficients $(α_u,~ α_l)$ for the upper and lower surfaces (not necessarily of the same lengths), trailing-edge displacement values $(Δz_u,~ Δz_l)$, and coefficients for leading edge modifications on the upper and lower surfaces at the nose.

AeroFuse.AircraftGeometry.leading_edge — Method

trailing_edge(wing :: Wing, flip = false)

Compute the trailing edge coordinates of a Wing, with an option to flip the signs of the $y$-coordinates.

AeroFuse.AircraftGeometry.lower_surface — Method

lower_surface(foil :: Foil)

Get the lower surface coordinates of a Foil from leading to trailing edge.

AeroFuse.AircraftGeometry.naca4 — Function

naca4(digits :: NTuple{4, <: Real}, n :: Integer = 40; sharp_trailing_edge :: Bool)

Generate a Foil of a NACA 4-digit series profile with a specified number of points (40 by default), and a named option to specify a sharp or blunt trailing edge.

AeroFuse.AircraftGeometry.naca4_coordinates — Method

naca4_coordinates(digits :: NTuple{4, <: Real}, n :: Integer, sharp_trailing_edge :: Bool)

Generate the coordinates of a NACA 4-digit series profile with a specified number of points, and a Boolean flag to specify a sharp or blunt trailing edge.

AeroFuse.AircraftGeometry.properties — Method

properties(wing :: AbstractWing)

Compute the generic properties of interest (span, area, etc.) of an AbstractWing.

AeroFuse.AircraftGeometry.read_foil — Method

read_foil(path :: String; header = true; name = "")

Generate a Foil from a file consisting of 2D coordinates with named arguments to skip the header (first line of the file) or assign a name.

By default, the header is assumed to exist and should contain the airfoil name, which is assigned to the name of the Foil.

AeroFuse.AircraftGeometry.scale — Method

scale(foil :: Foil, scale)

Scale the coordinates of a Foil to a scaling value.

AeroFuse.AircraftGeometry.split_surface — Method

split_surface(foil :: Foil)

Split the Foil coordinates into upper and lower surfaces.

AeroFuse.AircraftGeometry.surface_coordinates — Function

surface_coordinates(wing :: WingMesh, n_span = wing.num_span, n_chord = wing.num_chord)

Generate the surface coordinates of a WingMesh with default spanwise $n_s$ and chordwise $n_c$ panel distributions from the mesh.

AeroFuse.AircraftGeometry.surface_panels — Function

surface_panels(
    wing_mesh :: WingMesh, 
    n_s = wing_mesh.num_span, 
    n_c = length(first(foils(wing_mesh.surface))).x
)

Generate the surface panel distribution from a WingMesh with the default spanwise $n_s$ panel distribution from the mesh and the chordwise panel $n_c$ distribution from the number of points defining the root airfoil.

In case of strange results, provide a higher number of chordwise panels to represent the airfoils more accurately

AeroFuse.AircraftGeometry.sweeps — Function

sweeps(wing :: AbstractWing, w = 0.)

Obtain the sweep angles (in radians) at the corresponding normalized chord length ratio $w ∈ [0,1]$.

AeroFuse.AircraftGeometry.thickness_coordinates — Method

thickness_coordinates(coords :: Array{2, <: Real})

Generate the thickness coordinates on the $x$-$z$ plane at $y = 0$.

AeroFuse.AircraftGeometry.upper_surface — Method

upper_surface(foil :: Foil)

Get the upper surface coordinates of a Foil from leading to trailing edge.

AeroFuse.AircraftGeometry.wetted_area_ratio — Function

wetted_area_ratio(
    wing_mesh :: WingMesh, 
    n_s = wing_mesh.num_span, 
    n_c = length(first(foils(wing_mesh.surface))).x
)

Determine the wetted area ratio $S_{wet}/S$ of a WingMesh by calculating the ratio of the total area of the surface panels to the projected area of the Wing.

The wetted area ratio should be slightly above 2 for thin airfoils.

AeroFuse.AircraftGeometry.wing_bounds — Method

wing_bounds(wing :: Wing, flip = false)

Compute the leading and trailing edge coordinates of a Wing, with an option to flip the signs of the $y$-coordinates.

AeroFuse.PanelGeometry.make_panels — Method

make_panels(foil :: Foil)
make_panels(foil :: Foil, n :: Integer)

Generate a vector of Panel2Ds from a Foil, additionally with cosine interpolation using (approximately) $n$ points if provided.

AeroFuse.PanelGeometry.wetted_area — Function

wetted_area(
    wing_mesh :: WingMesh, 
    n_s = wing_mesh.num_span, 
    n_c = length(first(foils(wing_mesh.surface))).x
)

Determine the wetted area $S_{wet}$ of a WingMesh by calculating the total area of the surface panels.

Panel Geometry API

AeroFuse.PanelGeometry.Panel3D — Type

Panel3D(p1, p2, p3, p4)

Four Cartesian coordinates p1, p2, p3, p4 representing corners of a panel in 3 dimensions. The following commutative diagram (math joke) depicts the order:

z → y
↓
x
        p1 —→— p4
        |       |
        ↓       ↓
        |       |
        p2 —→— p3

AeroFuse.PanelGeometry.get_transformation — Function

get_transformation(panel :: AbstractPanel3D)

Generate the mapping to transform a point from global coordinates to an AbstractPanel3D's local coordinate system.

AeroFuse.PanelGeometry.make_panels — Method

make_panels(xyzs)

Convert an array of coordinates corresponding to a wing, ordered from root to tip and leading-edge to trailing-edge, into panels.

AeroFuse.PanelGeometry.midpoint — Method

midpoint(panel :: AbstractPanel3D)

Compute the midpoint of an AbstractPanel3D.

AeroFuse.PanelGeometry.normal_vector — Method

normal_vector(panel :: Panel3D)

Compute the normal vector of an AbstractPanel3D.

AeroFuse.PanelGeometry.panel_area — Method

panel_area(panel :: AbstractPanel3D)

Compute the area of a planar quadrilateral 3D panel.

AeroFuse.PanelGeometry.panel_coordinates — Method

panel_coordinates(panel :: Panel3D)

Compute the coordinates of a Panel3D.

AeroFuse.PanelGeometry.reflect_xz — Method

reflect_xz(panel :: Panel3D)

Reflect a Panel3D with respect to the $x$-$z$ plane of its reference coordinate system.

AeroFuse.PanelGeometry.transform — Method

transform(panel :: Panel3D, rotation, translation)

Perform an affine transformation on the coordinates of a Panel3D given a rotation matrix and translation vector.

AeroFuse.PanelGeometry.transform_normal — Method

transform_normal(panel :: Panel3D, h_l, g_l)

Transform the normal vector $n̂₀$ of a Panel3D about the hinge axis $ĥₗ$ by the control gain $gₗ$.

The transformation is the following: $n̂ₗ = gₗ ĥₗ × n̂₀$ (Flight Vehicle Aerodynamics, M. Drela, Eq. 6.36).

AeroFuse.PanelGeometry.wake_panel — Method

wake_panel(panels :: DenseArray{<: AbstractPanel3D}, bound, U)

Calculate required transformation from the global coordinate system to an to an AbstractPanel3D's local coordinate system.

AeroFuse.PanelGeometry.wetted_area — Method

wetted_area(panels :: Array{Panel3D})

Compute the total wetted area by summing the areas of an array of Panel3D.